+++ src/repos/index.ts
@@ -6,6 +6,7 @@ import { refreshRepositories } from "./store/actions";
 import { initializeStorage } from "./store/storage";
 import { registerTourController } from "./tours";
 import { registerTreeProvider } from "./tree";
+import { registerWikiController } from "./wiki";
 
 export function registerRepoModule(context: vscode.ExtensionContext) {
   registerRepoCommands(context);
@@ -17,6 +18,8 @@ export function registerRepoModule(context: vscode.ExtensionContext) {
 
   registerCommentController(context);
   registerTourController(context);
+
+  registerWikiController(context);
 
   refreshRepositories();
 }